+++ pubtools/pulplib/_impl/model/__init__.py
@@ -21,6 +21,7 @@ from .unit import (
     RpmUnit,
     RpmDependency,
     ModulemdUnit,
+    ModulemdDependency,
     ModulemdDefaultsUnit,
     YumRepoMetadataFileUnit,
 )